Transfer in parts of create-react-app's webpack config
