+++ webpack/utils.js
@@ -3,44 +3,27 @@ const CleanWebpackPlugin = require('clean-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const autoprefixer = require('autoprefixer');
 
-exports.setupCSS = function (paths) {
-  return {
-    module: {
-      loaders: [
-        {
-          test: /\.scss$/,
-          loaders: ['style', 'css', 'postcss', 'sass'],
-          include: paths,
-        },
-      ],
-    },
-    postcss: function () {
-      return [autoprefixer];
-    },
-  };
-}
-
 exports.minify = function () {
   return {
     plugins: [
       new webpack.optimize.UglifyJsPlugin({
-        // Don't beautify output (enable for neater output)
+        // Don't beautify output (enable for neater output).
         beautify: false,
-        // Eliminate comments
+        // Eliminate comments.
         comments: false,
-        // Compression specific options
+        // Compression specific options.
         compress: {
           warnings: false,
-          // Drop `console` statements
+          // Drop `console` statements.
           drop_console: true,
         },
-        // Mangling specific options
+        // Mangling specific options.
         mangle: {
-          // Don't mangle $
+          // Don't mangle $.
           except: ['$'],
-          // Don't care about IE8
+          // Don't care about IE8 because React doesn't support IE8.
           screw_ie8 : true,
-          // Don't mangle function names
+          // Don't mangle function names.
           keep_fnames: true,
         },
       }),
@@ -88,20 +71,52 @@ exports.clean = function (path) {
   };
 }
 
-exports.extractCSS = function (paths) {
+// CSS-related utils
+
+function postcss() {
+  return [
+    autoprefixer({
+      browsers: [
+        '>1%',
+        'last 4 versions',
+        'Firefox ESR',
+        'not ie < 9', // React doesn't support IE8 anyway.
+      ]
+    }),
+  ];
+}
+
+exports.setupCSS = function (paths) {
   return {
     module: {
       loaders: [
-        // Extract CSS during build
         {
-          test: /\.scss$/,
-          loader: ExtractTextPlugin.extract(['css', 'sass']),
+          test: /\.(css|scss)$/,
+          loaders: ['style', 'css', 'postcss', 'sass'],
           include: paths,
         },
       ],
     },
+    postcss: postcss,
+  };
+}
+
+exports.extractCSS = function (paths) {
+  return {
+    module: {
+      loaders: [
+        // Extract CSS during build.
+        {
+          test: /\.(css|scss)$/,
+          // Note: this won't work without `new ExtractTextPlugin()` in `plugins`.
+          loader: ExtractTextPlugin.extract(['css', 'postcss', 'sass']),
+          include: paths,
+        },
+      ],
+    },
+    postcss: postcss,
     plugins: [
-      // Output extracted CSS to a file
+      // Output extracted CSS to a file.
       new ExtractTextPlugin('[name].[chunkhash].css'),
     ],
   };
