+++ webpack/webpack.config.dev.js
@@ -1,15 +1,14 @@
 const merge = require('webpack-merge');
-const common = require('./webpack.common.config');
+const common = require('./webpack.config.common');
 const utils = require('./utils');
 const devServer = require('./dev-server');
 
-const environment = 'development';
 const config = merge(
   common,
   {
     devtool: 'eval-source-map',
   },
-  utils.setFreeVariable('process.env.NODE_ENV', environment),
+  utils.setFreeVariable('process.env.NODE_ENV', 'development'),
   utils.setupCSS(common.PATHS.styles),
   devServer({
     // Customize host/port here if needed
