+++ webpack/webpack.config.prod.js
@@ -1,8 +1,10 @@
 const merge = require('webpack-merge');
-const common = require('./webpack.common.config');
+const common = require('./webpack.config.common');
 const utils = require('./utils');
 
 // These dependencies will be extracted out into `vendor.js` in production build.
+// App bundle changes more often than vendor bundle and splitting app bundle from
+// 3rd-party vendor bundle allows the vendor bundle to be cached.
 const vendor = [
   'axios',
   'lodash',
@@ -15,12 +17,16 @@ const vendor = [
   'redux-thunk',
 ];
 
-const environment = 'production';
 const config = merge(
   common,
   {
+    // Don't attempt to continue if there are any errors.
+    bail: true,
+    // We generate sourcemaps in production. This is slow but gives good results.
+    // You can exclude the *.map files from the build during deployment.
     devtool: 'source-map',
     output: {
+      // The build folder.
       path: common.PATHS.build,
       filename: '[name].[chunkhash].js',
       // This is used for require.ensure. The setup
@@ -28,8 +34,10 @@ const config = merge(
       chunkFilename: '[chunkhash].js',
     },
   },
+  // Delete the build folder.
+  // TODO: Use create-react-app's way of building that shows file size differences.
   utils.clean(common.PATHS.build),
-  utils.setFreeVariable('process.env.NODE_ENV', environment),
+  utils.setFreeVariable('process.env.NODE_ENV', 'production'),
   utils.extractBundle({
     name: 'vendor',
     entries: vendor,
